gradle project can be imported from intelliJ
