+++ gradle.properties
@@ -1,11 +1,17 @@
-javacppPlatform=linux-x86_64,linux-armhf,linux-arm64,macosx-x86_64,macosx-arm64,windows-x86_64
-org.gradle.parallel=true
-
 outputName=ml4processing
 outputNamePostfix=
+
+versionNumber=1
+versionString=0.0.1
+
+# Uncomment to copy library to sketchbook directory
+sketchbookDirectory=/Users/andres/Documents/Processing
 
 # Uncomment for continues build with cuda:
 # cuda
 
 # Uncomment for no fat jar building
 disable-fatjar
+
+javacppPlatform=linux-x86_64,linux-armhf,linux-arm64,macosx-x86_64,macosx-arm64,windows-x86_64
+org.gradle.parallel=true
