+++ library.properties
@@ -4,7 +4,7 @@ authors=[Jeongin Lee](https://github.com/jjeongin)
 url=https://github.com/codeanticode/tensorflow-processing
 sentence=TensorFlow ML library for Processing
 paragraph=.
-version=1
-prettyVersion=0.0.1
+version=
+prettyVersion=
 minRevision=0
 maxRevision=0
